TB-13 Relocate components logic to hooks
