+++ src/shared/components/ListItem.tsx
@@ -1,49 +1,20 @@
-import { useDispatch, useSelector } from 'react-redux';
-import { tasksAPI } from '../api/tasksAPI';
+import { useListItem } from '../hooks/useListItem';
 import type { Task } from '../types/Task';
-import type { AppDispatch, RootState } from '../rtk/store';
-import { removeTask, setEditItem, updateTask } from '../rtk/tasksSlice';
 import cn from 'classnames';
-import { toast } from 'sonner';
-import { useState } from 'react';
 
 interface ListItemProps {
   task: Task;
 }
 
 export const ListItem = ({ task }: ListItemProps) => {
-  const dispatch = useDispatch<AppDispatch>();
-  const { editItem } = useSelector((state: RootState) => state.tasks);
-
-  const [isStatusChanging, setIsStatusChanging] = useState(false);
-  const [isDeletingTask, setIsDeletingTask] = useState(false);
-
-  const handleDeleteTask = async () => {
-    setIsDeletingTask(true);
-    await tasksAPI.remove(task.id);
-
-    toast.success(`Task '${task.title}' deleted.`);
-
-    dispatch(removeTask(task.id));
-    setIsDeletingTask(false);
-  };
-
-  const handleChengeStatusTask = async () => {
-    setIsStatusChanging(true);
-
-    const newTask = {
-      id: task.id,
-      title: task.title,
-      status: !task.status,
-      description: task.description,
-    };
-
-    await tasksAPI.update(task.id, newTask);
-
-    dispatch(updateTask(newTask));
-
-    setIsStatusChanging(false);
-  };
+  const {
+    editItem,
+    isStatusChanging,
+    isDeletingTask,
+    handleDeleteTask,
+    handleChengeStatusTask,
+    handleEditItem,
+  } = useListItem(task);
 
   return (
     <li
@@ -72,7 +43,7 @@ export const ListItem = ({ task }: ListItemProps) => {
       ) : (
         <div className='flex flex-col sm:flex-row gap-2 font-semibold'>
           <button
-            onClick={() => dispatch(setEditItem(task))}
+            onClick={handleEditItem}
             disabled={isStatusChanging || isDeletingTask}
           >
             Edit
